+++ src/index.js
@@ -4,7 +4,7 @@ import reducer from './reducers';
 
 export default class AdminUiPlugin extends Plugin {
   constructor() {
-    super('triniti', 'admin-ui', '0.4.13');
+    super('triniti', 'admin-ui', '0.4.14');
   }
 
   configure(app, bottle) {