+++ tests/components/modal/index.oldtest.jsx
@@ -11,7 +11,7 @@ let clock;
 let isOpenNested;
 let toggleNested;
 
-const setup = () => {
+const setup = (t) => {
   isOpen = false;
   toggle = () => { isOpen = !isOpen; };
 
@@ -19,7 +19,7 @@ const setup = () => {
   toggleNested = () => { isOpenNested = !isOpenNested; };
 
   clock = sinon.useFakeTimers();
-};
+}
 
 const teardown = () => {
   clock.tick(300);
@@ -34,76 +34,46 @@ const test = (description, fn) => {
   });
 };
 
-test('Modal:: should render modal portal into DOM', (t) => {
-  isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle}>
-      Yo!
-    </Modal>
-  ));
-
-  clock.tick(300);
-  t.notEqual(wrapper.childAt(0).children().length, 0);
-  wrapper.unmount();
-
-  t.end();
-});
-
 test('Modal:: should render with the class "modal-dialog"', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 1);
   wrapper.unmount();
-
-  t.end();
-});
-
-test('Modal:: should render external content when present', (t) => {
-  isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle} external={<button className="cool-close-button">&times;</button>}>
-      Yo!
-    </Modal>
-  ));
-
-  clock.tick(300);
-  t.equal(document.getElementsByClassName('cool-close-button').length, 1);
-  t.equal(document.getElementsByClassName('cool-close-button')[0].nextElementSibling.className.split('Modal::  ').indexOf('modal-dialog') > -1, true);
-  wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should render with the backdrop with the class "modal-backdrop" by default', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 1);
   wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should render with the backdrop with the class "modal-backdrop" when backdrop is "static"', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} backdrop="static">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 1);
   wrapper.unmount();
   t.end();
@@ -111,13 +81,14 @@ test('Modal:: should render with the backdrop with the class "modal-backdrop" wh
 
 test('Modal:: should not render with the backdrop with the class "modal-backdrop" when backdrop is "false"', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} backdrop={false}>
       Yo!
     </Modal>
-  ));
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 1);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 0);
   wrapper.unmount();
@@ -126,82 +97,68 @@ test('Modal:: should not render with the backdrop with the class "modal-backdrop
 
 test('Modal:: should render with class "modal-dialog" and have custom class name if provided', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} className="my-custom-modal">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 1);
   t.equal(document.getElementsByClassName('my-custom-modal').length, 1);
   wrapper.unmount();
   t.end();
 });
 
-test('Modal:: should render with class "modal-dialog" w/o centered class if not provided', (t) => {
+test('Modal:: should render with maxWidth prop', (t) => {
   isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle}>
+  const wrapper = mount(
+    <Modal isOpen={isOpen} toggle={toggle} maxWidth="95%" >
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
-  t.equal(document.getElementsByClassName('modal-dialog').length, 1);
-  t.equal(document.getElementsByClassName('modal-dialog-centered').length, 0);
-  wrapper.unmount();
-  t.end();
-});
-
-test('Modal:: should render with class "modal-dialog" and centered class if provided', (t) => {
-  isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle} centered>
-      Yo!
-    </Modal>
-  ));
-
-  clock.tick(300);
-  t.equal(document.getElementsByClassName('modal-dialog').length, 1);
-  t.equal(document.getElementsByClassName('modal-dialog-centered').length, 1);
+  t.equal(document.getElementsByClassName('modal-dialog')[0].style.maxWidth, '95%');
   wrapper.unmount();
   t.end();
 });
 
 test('Modal:: should render with additional props if provided', (t) => {
   isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle} style={{ maxWidth: '95%' }}>
+  const wrapper = mount(
+    <Modal isOpen={isOpen} toggle={toggle} style={{ height: '95%' }}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 1);
-  t.equal(document.getElementsByClassName('modal-dialog')[0].style.maxWidth, '95%');
+  t.equal(document.getElementsByClassName('modal-dialog')[0].style.height, '95%', document.getElementsByClassName('modal-dialog')[0].style.height);
   wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should render without fade transition if provided with fade={false}', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} fade={false} modalClassName="fadeless-modal">
       Howdy!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   // Modal should appear instantaneously
   clock.tick(1);
+  t.equal(wrapper.children().length, 0);
 
   const matchedModals = document.getElementsByClassName('fadeless-modal');
   const matchedModal = matchedModals[0];
 
   t.equal(matchedModals.length, 1);
   // Modal should not have the 'fade' class
-  t.true(matchedModal.className.split('Modal::  ').indexOf('fade') < 0);
+  t.true(matchedModal.className.split(' ').indexOf('fade') < 0);
 
   wrapper.unmount();
   t.end();
@@ -209,7 +166,7 @@ test('Modal:: should render without fade transition if provided with fade={false
 
 test('Modal:: should render when expected when passed modalTransition and backdropTransition props', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal
       isOpen={isOpen}
       toggle={toggle}
@@ -218,10 +175,11 @@ test('Modal:: should render when expected when passed modalTransition and backdr
       modalClassName="custom-timeout-modal"
     >
       Hello, world!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(20);
+  t.equal(wrapper.children().length, 0);
 
   const matchedModals = document.getElementsByClassName('custom-timeout-modal');
 
@@ -233,13 +191,14 @@ test('Modal:: should render when expected when passed modalTransition and backdr
 
 test('Modal:: should render with class "modal" and have custom class name if provided with modalClassName', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} modalClassName="my-custom-modal">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.querySelectorAll('.modal.my-custom-modal').length, 1);
   wrapper.unmount();
   t.end();
@@ -247,13 +206,14 @@ test('Modal:: should render with class "modal" and have custom class name if pro
 
 test('Modal:: should render with custom class name if provided with wrapClassName', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} wrapClassName="my-custom-modal">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('my-custom-modal').length, 1);
   wrapper.unmount();
   t.end();
@@ -261,13 +221,14 @@ test('Modal:: should render with custom class name if provided with wrapClassNam
 
 test('Modal:: should render with class "modal-content" and have custom class name if provided with contentClassName', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} contentClassName="my-custom-modal">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.querySelectorAll('.modal-content.my-custom-modal').length, 1);
   wrapper.unmount();
   t.end();
@@ -275,13 +236,14 @@ test('Modal:: should render with class "modal-content" and have custom class nam
 
 test('Modal:: should render with class "modal-backdrop" and have custom class name if provided with backdropClassName', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} backdropClassName="my-custom-modal">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.querySelectorAll('.modal-backdrop.my-custom-modal').length, 1);
   wrapper.unmount();
   t.end();
@@ -289,29 +251,30 @@ test('Modal:: should render with class "modal-backdrop" and have custom class na
 
 test('Modal:: should render with the class "modal-${size}" when size is passed', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} size="crazy">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 1);
   t.equal(document.getElementsByClassName('modal-crazy').length, 1);
   wrapper.unmount();
   t.end();
 });
 
-
 test('Modal:: should render modal when isOpen is true', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal').length, 1);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 1);
   wrapper.unmount();
@@ -320,13 +283,14 @@ test('Modal:: should render modal when isOpen is true', (t) => {
 
 test('Modal:: should render modal with default role of "dialog"', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal')[0].getAttribute('role'), 'dialog');
   wrapper.unmount();
   t.end();
@@ -334,60 +298,58 @@ test('Modal:: should render modal with default role of "dialog"', (t) => {
 
 test('Modal:: should render modal with provided role', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} role="alert">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal')[0].getAttribute('role'), 'alert');
   wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should render modal with aria-labelledby provided labelledBy', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} labelledBy="myModalTitle">
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal')[0].getAttribute('aria-labelledby'), 'myModalTitle');
   wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should not render modal when isOpen is false', (t) => {
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
-  t.equal(wrapper.childAt(0).children().length, 0);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal').length, 0);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 0);
   wrapper.unmount();
-
   t.end();
 });
 
 test('Modal:: should toggle modal', (t) => {
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
   t.false(isOpen);
-  t.equal(wrapper.childAt(0).children().length, 0);
   t.equal(document.getElementsByClassName('modal').length, 0);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 0);
 
@@ -399,7 +361,6 @@ test('Modal:: should toggle modal', (t) => {
   t.equal(document.getElementsByClassName('modal').length, 1);
   t.equal(document.getElementsByClassName('modal-backdrop').length, 1);
   wrapper.unmount();
-
   t.end();
 });
 
@@ -408,11 +369,11 @@ test('Modal:: should call onClosed & onOpened', (t) => {
   const onClosed = sinon.spy(Modal.prototype, 'onClosed');
   const handleOpen = sinon.spy();
   const handleClose = sinon.spy();
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} onOpened={handleOpen} onClosed={handleClose} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
   t.false(isOpen);
@@ -461,7 +422,7 @@ test('Modal:: should call onClosed & onOpened when fade={false}', (t) => {
   wrapper.setProps({ isOpen });
   clock.tick(1);
 
-  t.true(isOpen);
+  t.true(isOpen)
   t.true(onOpened.called);
   t.true(onClosed.notCalled);
 
@@ -476,38 +437,68 @@ test('Modal:: should call onClosed & onOpened when fade={false}', (t) => {
   t.end();
 });
 
-test('Modal:: should not call init when isOpen does not change', (t) => {
-  const init = sinon.spy(Modal.prototype, 'init');
+test('Modal:: should not call togglePortal when isOpen does not change', (t) => {
+  const togglePortal = sinon.spy(Modal.prototype, 'togglePortal');
   const componentDidUpdate = sinon.spy(Modal.prototype, 'componentDidUpdate');
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
   t.false(isOpen);
-  t.true(init.notCalled);
+  t.true(togglePortal.notCalled);
   t.true(componentDidUpdate.notCalled);
 
   wrapper.setProps({ isOpen });
   clock.tick(300);
 
   t.false(isOpen);
-  t.true(init.notCalled);
+  t.true(togglePortal.notCalled);
   t.true(componentDidUpdate.called);
 
+  togglePortal.restore();
+  componentDidUpdate.restore();
+  wrapper.unmount();
+  t.end();
+});
+
+test('Modal:: should renderIntoSubtree when props updated', (t) => {
+  isOpen = true;
+  const togglePortal = sinon.spy(Modal.prototype, 'togglePortal');
+  const renderIntoSubtree = sinon.spy(Modal.prototype, 'renderIntoSubtree');
+  const wrapper = mount(
+    <Modal isOpen={isOpen} toggle={toggle}>
+      Yo!
+    </Modal>,
+  );
+
+  clock.tick(300);
+  t.true(isOpen);
+  t.equal(togglePortal.callCount, 1);
+  t.equal(renderIntoSubtree.callCount, 1);
+
+  wrapper.setProps({ isOpen });
+  clock.tick(300);
+
+  t.true(isOpen);
+  t.equal(togglePortal.callCount, 1);
+  t.equal(renderIntoSubtree.callCount, 2);
+
+  togglePortal.restore();
+  renderIntoSubtree.restore();
   wrapper.unmount();
   t.end();
 });
 
 test('Modal:: should close modal when escape key pressed', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
   const instance = wrapper.instance();
 
   clock.tick(300);
@@ -527,7 +518,7 @@ test('Modal:: should close modal when escape key pressed', (t) => {
   t.false(isOpen);
 
   wrapper.setProps({ isOpen });
-  clock.tick(300);
+  clock.tick(301);
 
   t.equal(document.getElementsByClassName('modal').length, 0);
 
@@ -537,11 +528,11 @@ test('Modal:: should close modal when escape key pressed', (t) => {
 
 test('Modal:: should not close modal when escape key pressed when keyboard is false', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} keyboard={false}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
   const instance = wrapper.instance();
 
   clock.tick(300);
@@ -571,20 +562,19 @@ test('Modal:: should not close modal when escape key pressed when keyboard is fa
 
 test('Modal:: should close modal when clicking backdrop', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       <button id="clicker">Does Nothing</button>
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
 
   t.true(isOpen);
   t.equal(document.getElementsByClassName('modal').length, 1);
-  //
+
   document.getElementById('clicker').click();
   clock.tick(300);
-
   t.true(isOpen);
 
   document.getElementsByClassName('modal')[0].click();
@@ -598,11 +588,11 @@ test('Modal:: should close modal when clicking backdrop', (t) => {
 
 test('Modal:: should not close modal when clicking backdrop and backdrop is "static"', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle} backdrop="static">
       <button id="clicker">Does Nothing</button>
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
 
@@ -625,67 +615,47 @@ test('Modal:: should not close modal when clicking backdrop and backdrop is "sta
 
 test('Modal:: should destroy this._element', (t) => {
   isOpen = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       <button id="clicker">Does Nothing</button>
-    </Modal>
-  ));
+    </Modal>,
+  );
   const instance = wrapper.instance();
 
   clock.tick(300);
   t.ok(instance._element);
 
-  toggle();
-  wrapper.setProps({ isOpen });
-  clock.tick(300);
+  instance.destroy();
 
-  t.false(isOpen);
-  t.equal(instance._element, null);
+  t.same(instance._element, null);
 
+  instance.destroy();
   wrapper.unmount();
-  t.end();
-});
-
-test('Modal:: should destroy this._element on unmount', (t) => {
-  isOpen = true;
-  const wrapper = mount((
-    <Modal isOpen={isOpen} toggle={toggle}>
-      <button id="clicker">Does Nothing</button>
-    </Modal>
-  ));
-  const instance = wrapper.instance();
-
-  clock.tick(300);
-  t.ok(instance._element, 'instance._element should be truthy');
-
-  wrapper.unmount();
-  clock.tick(300);
-
-  t.equal(instance._element, null);
-
   t.end();
 });
 
 test('Modal:: should render nested modals', (t) => {
   isOpen = true;
   isOpenNested = true;
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       <ModalBody>
         <Modal isOpen={isOpenNested} toggle={toggleNested}>
           Yo!
         </Modal>
       </ModalBody>
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   clock.tick(300);
+  t.equal(wrapper.children().length, 0);
   t.equal(document.getElementsByClassName('modal-dialog').length, 2);
   t.equal(document.body.className, 'modal-open modal-open');
 
   wrapper.unmount();
   t.equal(document.getElementsByClassName('modal-dialog').length, 0);
   t.equal(document.body.className, '');
+
   t.end();
 });
 
@@ -693,11 +663,11 @@ test('Modal:: should remove exactly modal-open class from body', (t) => {
   // set a body class which includes modal-open
   document.body.className = 'my-modal-opened';
 
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   // assert that the modal is closed and the body class is what was set initially
   clock.tick(300);
@@ -732,11 +702,11 @@ test('Modal:: should remove exactly modal-open class from body', (t) => {
 test('Modal:: should call onEnter & onExit props if provided', (t) => {
   const onEnter = sinon.spy();
   const onExit = sinon.spy();
-  const wrapper = mount((
+  const wrapper = mount(
     <Modal isOpen={isOpen} onEnter={onEnter} onExit={onExit} toggle={toggle}>
       Yo!
-    </Modal>
-  ));
+    </Modal>,
+  );
 
   t.false(isOpen);
   t.true(onEnter.called);
@@ -763,19 +733,6 @@ test('Modal:: should call onEnter & onExit props if provided', (t) => {
   wrapper.unmount();
   t.true(onEnter.notCalled);
   t.true(onExit);
-
-  t.end();
-});
-
-test('Modal:: should update element z index when prop changes', (t) => {
-  const wrapper = shallow((
-    <Modal isOpen zIndex={0}>
-      Yo!
-    </Modal>
-  ));
-  t.equal(wrapper.instance()._element.style.zIndex, '0');
-  wrapper.setProps({ zIndex: 1 });
-  t.equal(wrapper.instance()._element.style.zIndex, '1');
 
   t.end();
 });